+++ src/default_branch.rs
@@ -1,6 +1,5 @@
-use crate::api;
-use crate::api::RemoteRepo;
-use crate::rest_api;
+use crate::github;
+use crate::github::{NoReposFound, RemoteRepo, Unauthorized};
 
 use anyhow::{Context, Result};
 
@@ -42,7 +41,7 @@ impl DefaultBranchArgs {
 }
 
 fn set_default_branch(repo: &RemoteRepo, default_branch: &str, token: &str) -> Result<()> {
-    rest_api::set_default_branch(repo, default_branch, token)
+    github::set_default_branch(repo, default_branch, token)
 }
 
 fn get_user_token() -> Result<String> {
@@ -51,13 +50,13 @@ fn get_user_token() -> Result<String> {
 }
 
 fn get_remote_repos(token: &str, org: &str) -> Result<Vec<RemoteRepo>> {
-    match api::list_org_repos(token, org).context("Fetching repositories") {
+    match github::list_org_repos(token, org).context("Fetching repositories") {
         Ok(repos) => Ok(repos),
         Err(e) => {
-            if let Some(_) = e.downcast_ref::<crate::api::NoReposFound>() {
+            if let Some(_) = e.downcast_ref::<NoReposFound>() {
                 anyhow::bail!("No repositories found");
             }
-            if let Some(_) = e.downcast_ref::<crate::api::Unauthorized>() {
+            if let Some(_) = e.downcast_ref::<Unauthorized>() {
                 anyhow::bail!("User token invalid. Run dadmin init with a valid token");
             }
             return Err(e);
